Port TestChronoZonedDateTime to scalatest
